Add support for R7RS's "indented string" escape syntax.
This allows the user to "escape" all whitespace surrounding *one*
newline with a backslash, causing all of this to be completely
collapsed to nothing.  This is useful when writing long string
literals which should be broken up into multiple lines.

This also adds some tests for other R7RS escape syntax, most of which
we already supported anyway.  There's only the escaped hex scalar
value left, which is ambiguous with regard to the old CHICKEN hex
scalar syntax.